requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
pillow==10.4.0
sympy==1.13.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
